Refactored components/ and pages/ folders
