+++ src/components/Core/DropzoneBox.tsx
@@ -7,49 +7,21 @@ import { useDropzone } from "react-dropzone";
 // Types
 import type { DropzoneOptions } from "react-dropzone";
 
-function readFile(file: File): Promise<string> {
-  return new Promise((resolve, reject) => {
-    const reader: FileReader = new FileReader();
-
-    // Define event handlers
-    reader.onabort = () => console.log("File reading aborted!");
-    reader.onerror = () => console.log("File reading failed!");
-    reader.onload = () => resolve(reader.result as string);
-    
-    // reader.readAsArrayBuffer(file); // Read given file as ArrayBuffer
-    // reader.readAsBinaryString(file);  // Read given file as binary
-    reader.readAsDataURL(file); // Read given file as Base64 String
-  });
-}
-
 export default function DropzoneBox({
   children,
-  setGallery,
+  handleOnDrop,
+  maxFiles,
 }: any) {
 
   // react-dropzone
   const options: DropzoneOptions = {
-    maxFiles: 5,
+    maxFiles: maxFiles,
     accept: {
       "image/*": [],
     },
-    onDrop: (acceptedFiles: File[]) => {
-      async function init() {
-        const gallery_new: any[] = [];
-        for (let i = 0; i < acceptedFiles.length; i++) {
-          const file: any = acceptedFiles[i];
-          const file_new: any = {
-            name: file.path,
-            preview: URL.createObjectURL(file),
-            data: await readFile(file),
-            type: file.type,
-          };
-          gallery_new.push(file_new);
-        }
-        setGallery(gallery_new);
-      }
-
-      init();
+    onDrop: (files: File[]) => {
+      if (files.length > 0) handleOnDrop(files);
+      else console.error("Error with Gallery input: exceeded 5-image maximum!")
     },
   };
   const { getRootProps, getInputProps } = useDropzone(options);
