+++ src/components/CreateLAG/Controls.tsx
@@ -1,6 +1,6 @@
 import wait from "../../utils/wait";
 import { Flex } from "@chakra-ui/react";
-import ToggleBtn from "../Core/ToggleBtn";
+import ToggleBtn from "./ToggleBtn";
 import { pressBtn } from "../../utils/animations";
 import { useState, useEffect, ReactElement } from "react";
 import CurveSubContainer from "../Core/CurveSubContainer";
