+++ src/pages/discover.tsx
